python_assignment_e-commerce/ecommerce_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose[cryptography]
python-dotenv
psycopg2-binary
alembic
passlib
fastapi-mail
